bulk update to push to GH
